Remove tChannelID::ClrPolarization() and optional parameters for cChannelManager::GetByChannelID()
ClrPolarization() seems to be part of a hack where cChannels use the TID
field of their channel ID to store frequency (HMz). An unrelated hack adds
100GHz-400GHz for different polarizations. Which means ClrPolarization() is
a hack for a hack. Good riddance.

Possible side effect: The only use of ClrPolarization() was in cTimer (see
commit), which means it no longer applies the polarization hack when
searching for its channel.

This continues cleanup performed in parent commit 5b854d2.