Fix section filters being delivered the wrong table
The error was introduced shortly after the receiver refactor, when duplicate
functionality in the section filter & receiver subsystems was merged in
a3cf51a16.

Section filter and receiver resources were abstracted as a single resource
type defined by a PID (see devices/PIDResource.h). However, this new
abstraction over-simplified section filter resources, as only PID was used
to match resources and the TID/mask was ignored. As a result, the TVCT table
(PID=0x1FFB, TID=0xC8) was being given to the MGT filter (which requested
PID=0x1FFB, TID=0xC7).

The solution is to move the comparison function into the resource
abstraction so that it may take into account its additional parameters.

This commit also makes opening resources the responsibility of the base
device subsystem (which was originally intended, as the Open() function
appears in the base abstraction).