Fix Channel(), Scan() and cScanner having different device types
This makes Scan() rely on Channel() for device type identification instead
of relying on its own (possibly inconsistent) idea of what type the device
is. Additionally, cScanner determines device type from Channel() instead of
allowing the user to override a (possibly inconsistent) type.

This fixes cScanner deciding that the device is terrestrial instead of ATSC,
and the resulting confusion when it forces Scan()'s type to terrestrial. The
problem manifested when cScanner set Scan()'s type to terrestrial (even
though scan type was forced to ATSC), causing EIT filters to be attached to
the device, segfaulting when our dear friend libsi attempted to parse
sections masquerading as EIT tables.

This bug masked another potential problem in the code, now made obvious in
cScanner; a device can provide multiple types, and cScanner chooses the
first available one. This could be fixed by allowing the user to choose a
source when multiple are provided, or simply scanning both types. Scanning
multiple types could be a lengthy process, so multiple scans could be done
in parallel until the first channel is discovered, discontinuing the other
types.